+++ R/utils.R
@@ -84,3 +84,14 @@ captureDots <- function() {
   args <- pairlist(parent.frame())
   .Call(rlang_capturedots, NULL, NULL, args, environment())
 }
+
+meow <- function(...) {
+  cat(chr_lines(..., .trailing = TRUE))
+}
+chr_lines <- function(..., .trailing = FALSE) {
+  lines <- paste(chr(...), collapse = "\n")
+  if (.trailing) {
+    lines <- paste0(lines, "\n")
+  }
+  lines
+}